+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Additional Changes:
 * WEB-1631: `Layer.Core.Message.deliveryStatus` and `Layer.Core.Message.readStatus` now more correctly handles case where `Layer.Core.Conversation` is still loading
   from the server.
 * WEB-1421: Client now throws errors if a `challenge` event is not registered.  All apps should register this event even if only to handle reauthentication
+* WEB-1198: Filters out invalid Identities from `createConversation` and `addParticipants`
+
 
 ## 1.0.0-pre2.8